refactor(console): 拆分经验治理 helper 与列定义
- 抽出经验治理格式化、类型和规则预设 helper

- 抽出等级、统计、流水和治理留痕表格列

- 保持经验规则、冻结语义和数据契约不变
